Improve tests and move some fo the link_editor
